Fix double declaration of yylineno
The yylineno variable was declared in dt_cc.c *and* in dt_lex.c (which
is generated from dt_lex.l).  This poses a compilation issue with some
newer compilers.  The declaration in dt_cc,c has been changed into an
extern declaration, while the one in dt_lex.c is kept as the actual
variable declaration.

Signed-off-by: Kris Van Hees <kris.van.hees@oracle.com>