Record CPU that BEGIN and END probes run on in the 'state' BPF map
In order to ensure that the BEGIN probe trace data is processed first,
and END probe trace data is processed last, we need to know what CPU
buffers their data is stored in.  This patch adds code to record the
CPU on which the BEGIN and END probes run.  The 'state' BPF map has
entries (DT_STATE_BEGANON and DT_STATE_ENDEDON) to store this data.

Signed-off-by: Kris Van Hees <kris.van.hees@oracle.com>
Reviewed-by: Eugene Loh <eugene.loh@oracle.com>
Reviewed-by: Kris Van Hees <kris.van.hees@oracle.com>