Move BPF stack, dctx, and mstate definitions out of dt_impl.h
The BPF stack layout defines do not belong in dt_impl.h and therefore
are moved to dt_dctx.h.  The dt_impl.h does not need to include
dt_dctx.h either because many files do not need it.

This change has a lot of cascading effects across various source code
files due to the include file content reschuffling, so a lot of small
tweaks are included in this patch.

Signed-off-by: Kris Van Hees <kris.van.hees@oracle.com>
Reviewed-by: Eugene Loh <eugene.loh@oracle.com>
Reviewed-by: Kris Van Hees <kris.van.hees@oracle.com>