Use default action ONLY if the clause is empty
If a clause is empty, use the default clause.  This was working.

On the other hand, if a clause is not empty, do not use the default
clause -- even if the clause traces no data!  This was not working
in DTv2.  Arguably, the new behavior was more sensible, but it
breaks precedent, including in some very important and common
cases, such as clauses that simply manipulate D variables.
Restore the legacy behavior.

Do not modify the prologue.  A few instructions in the prologue could
be eliminated for clauses that will not write the output buffer, but
the instructions are inconsequential and would be difficult to eliminate.

The test test/unittest/funcs/tst.default.d was too mild to catch
this problem.  Add test/unittest/actions/default/tst.default.d .

Orabug: 31772287
Signed-off-by: Eugene Loh <eugene.loh@oracle.com>
Reviewed-by: Kris Van Hees <kris.van.hees@oracle.com>