Adjust xfail annotations due to line-number change
An earlier patch noted that some .r test-result files checked line
numbers.  So, once an @@xfail annotation was added to the test D script,
line numbers in the script were displaced and it became impossible for
the test to pass.  That patch identified a few cases where simply
eliminating the @@xfail annotation caused the test to pass again.

Search methodically for all cases where the .r file checks line numbers
and the D script has an @@xfail annotation.  Restore line numbers in the
script, either by removing the annotation or making it a substitution
rather than an insertion.

Signed-off-by: Eugene Loh <eugene.loh@oracle.com>
Reviewed-by: Kris Van Hees <kris.van.hees@oracle.com>