Implement a 'aggs' per-CPU BPF map to store aggregation data
Aggregation data consists of data items (all uint64_t) that are updated
when an aggregation function is executed.  This is quite different from
other data recording actions that append data to an output buffer.

We need support for concurrency control to allow reading the data during
non-atomic updates.  We use a latch mechanism, i.e. a multiversion
concurrency control mechanism, to satify these requirements.

Simply put, the per-CPU aggregation data area stores a sequence id that
tracks the generation of the data updates.  Since only one aggregation
can be updated at a time on a single CPU, one latch is sufficient for
all aggregations (per CPU-buffer).  We also allocate two copies of each
aggregation, which we will call A and B.

When we update the aggregation data, we first modify A and then we modify
B, while ensuring that while we modify one, the reader is directed to
read from the other.

Signed-off-by: Kris Van Hees <kris.van.hees@oracle.com>
Reviewed-by: Eugene Loh <eugene.loh@oracle.com>