test: tst.ustackdepth-bug25949692.d timeout
The test would timeout on some systems -- e.g., ARM and x86 VMs
running OL8.  Since it relied on syscall::ioctl:entry firing but there
was no trigger, there was no guarantee that the test would complete in
any particular time.

Remove the test.  Its scope was very narrow (testing only that
ustackdepth was neither 1 nor 0, in response to a particular bug)
and focused on behavior in an implementation that is no longer used.
Meanwhile, this built-in variable is tested more rigorously by other
tests that, e.g., check the ustackdepth value against the actual
depth of the ustack.

Orabug: 31929511
Signed-off-by: Eugene Loh <eugene.loh@oracle.com>
Reviewed-by: Kris Van Hees <kris.van.hees@oracle.com>