Allocate exit label in dt_cg_tramp_prologue rather than each provider
Every provider implementation was allocating an id for the exit label
that is used in the trampoline prologue and epilogue.  Now the label
id is allocated in and returned by dt_cg_tramp_prologue() (and its
*_act() variant).  That id is to be passed to dt_cg_tramp_epilogue()
(or the *_advance() variant) to ensure that the same label is used
throughout the trampoline.

Signed-off-by: Kris Van Hees <kris.van.hees@oracle.com>
Reviewed-by: Eugene Loh <eugene.loh@oracle.com>
Reviewed-by: Kris Van Hees <kris.van.hees@oracle.com>