Ensure stable testing criteria for profile-n basic tests
The test/unittest/profile-n/tst.profile*.d test cases were written
to generate output that was compared with expected output.  They all
followed the same basic template:

	profile-<spec>
	{
		printf(<string>);
		exit(0);
	}

This will trigger a probe that fires on all CPUs at roughly the same
time.  This means that <string> may be written to the output buffer of
multiple CPUs.  Due to a bug in the exit() action implementation on the
reference DTrace implementation, only the trace data (<string> in this
case) of the CPU that executes it sexit(0) statement first would be
displayed as output while the data for other CPUs was silently dropped.

The current DTrace implementation (based on BPF) does not exhibit this
buggy behaviour and therefore multiple instances of <string> may be
reported in output.

Given that these tests merely verify that the profile-<spec> probe name
is valid and results in a successful probing session, there is no need
to output any data.  If the dtrace command completes successfully with
return code 0, the test succeeded.

The tst.basic.d test (and result file) have been removed because the
test is a duplicate of tst.profilenhz.d.

Signed-off-by: Kris Van Hees <kris.van.hees@oracle.com>
Reviewed-by: Eugene Loh <eugene.loh@oracle.com>
Reviewed-by: Kris Van Hees <kris.van.hees@oracle.com>