Ensure the BPF gcc compiler is new enough
Some early releases of the BPF support in gcc generated code in the
function prologue to save registers %r6 and on (based on need).  The
BPF verifier in the kernel does not allow loading from registers that
were never stored to.  The instructions to save the content of %r6
and on caused the BPF verifier to reject the program.

We now perform a test compile of a BPF function that is known to
require enough registers to ensure this failure mode is exercised.

Signed-off-by: Kris Van Hees <kris.van.hees@oracle.com>