Ensure that bash is used as shell for building
The makefiles depend on POSIX features such as [[ ... ]] that are not
available in the default shell on some distributions like Debian and
Ubuntu.  Explicitly require bash to be used as shell.

This also reverts a previous patch that replaced the [[ ... ]] construct
with the use of the 'test' command (4506ce52 "Support building when shell
is sh.").

Signed-off-by: Kris Van Hees <kris.van.hees@oracle.com>