END clause does not execute when consumer stops the producer
The END probe's clause does not execute when the consumer requests
tracing to stop.  One example is when the user sends a SIGINT to the
dtrace command-line tool and it calls dtrace_stop().  This problem
would also occur once -c is supported and the specified command finishes.

The problem is that the END probe checks that the activity state is
DRAINING before executing its clause, but DRAINING is set only by
calling the D exit() action.

Before calling END_probe(), advance the activity state to DRAINING if
it is not yet that far.  Add a test for this problem.

Orabug: 32036406
Signed-off-by: Eugene Loh <eugene.loh@oracle.com>
Reviewed-by: Kris Van Hees <kris.van.hees@oracle.com>