sdt and syscall probe descriptions broken when multiple probes
The sdt and syscall providers are broken in that, when a probe
description corresponds to multiple probes, a BPF program is
attached to only one of the probes.  E.g., syscall:::entry will
attach a BPF program to only one probe, even though hundreds of
probes are indicated!  The problem is that the attach function
looks for an event_id, skipping the probe if there is no such ID,
but the ID exists for only the "representative probe."

Therefore, call dt_probe_info() when we append probes to the
dt_enablings list.

What is surprising is that, with this fix, there are no XPASS
tests.  That is, there had presumably been no XFAIL tests
("expected fails") that had been failing due *only* to this
problem.  A quick survey of tests suggests that this is likely;
remarkably few tests rely on such functionality, and those that
do still fail due to other missing support (such as strings or
other providers).

Therefore, add a test to check the fix.

Orabug: 32036431
Signed-off-by: Eugene Loh <eugene.loh@oracle.com>
Reviewed-by: Kris Van Hees <kris.van.hees@oracle.com>