No data is reported for quantize() when it's all in the last bin
The search for a nonzero bin is broken for quantize(), causing
no data to be reported when it's all in the last bin.

Fix the search for the "first bin".  Make a few other small
changes so that the three *quantize() functions are more similar
and such coding errors are more apparent.

Add a test to catch this error.  The test is XFAIL for now,
because we do not have a consumer yet.

Orabug: 32148161
Signed-off-by: Eugene Loh <eugene.loh@oracle.com>
Reviewed-by: Kris Van Hees <kris.van.hees@oracle.com>