Clean up agg identifier storage data handling
The storage layout for aggregations is based on storing offset and
size data in the identifier.  This will now be done using the new
dt_ident_set_storage() function.  The dt_ident_t structure gets two
new fields:
- di_size: storage size of the identifier (only used for aggregations)
- di_hash: idhash that the identifier belongs to

This patch also introduces a DT_CG_AGG_SET_STORAGE() macro to handle
the code for only setting the storage data the first time we encounter
an aggregation.

Signed-off-by: Kris Van Hees <kris.van.hees@oracle.com>
Reviewed-by: Eugene Loh <eugene.loh@oracle.com>