Implement aggregation functions: max(), min(), sum(), and stddev()
The min() and max() functions require initialization of their initial
value.  That is, data for min() is initialized with the highest
possible value and data for max() is initialized with the lowest
possible value.  This initialization happens in dt_aggregate_go().

The stddev() function records the number of values, the sum of values,
and the 128-bit sum of the squares of each value.  This information is
processed by the consumer to determine the standard deviation value.

Signed-off-by: David Mc Lean <david.mclean@oracle.com>
Signed-off-by: Kris Van Hees <kris.van.hees@oracle.com>
Reviewed-by: Kris Van Hees <kris.van.hees@oracle.com>