Add a multi-CPU aggregation test
A key characteristic of DTrace aggregations is that they are
gathered per CPU and then aggregated across CPUs.  Nevertheless,
the test suite uses only BEGIN and tick-n probes for aggregation
tests.  So, aggregation across CPUs is never really tested.

Add a test for the simplest aggregations with a probe that fires
across all CPUs.

Signed-off-by: Eugene Loh <eugene.loh@oracle.com>
Reviewed-by: Kris Van Hees <kris.van.hees@oracle.com>