Test min/max aggregation initialization
The min and max aggregation functions have a special code path in
which initial values are set to max and min values, respectively.
All other aggregation functions expect zeroes for initialization.
The min and max tests include zeroes as values, and therefore do
not detect if the special code path is not used.

Change the min and max tests to use values of a particular sign
(positive by default, negative in the _neg variants) to exercise
the special code path.

Signed-off-by: Eugene Loh <eugene.loh@oracle.com>
Reviewed-by: Kris Van Hees <kris.van.hees@oracle.com>