ERROR probe implementation
The ERROR probe is special because it is not associated with a system
event.  Instead, it is a 'fake' probe that is reported whenever there
a fault occurs during probe execution.  This can be implemented in
BPF by including the ERROR probe clauses in every probe program that
needs them.

Whenever a fault occurs in one of our BPF programs, a call is made to
dt_probe_error() to record data associated with the fault, and the
clause is terminated with a non-zero return value.  The data items
that can be set are:

	- offset in the clause at which the fault occurred
	- fault type
	- value

When a clause ends with a non-zero return value, execution is directed
to the ERROR probe clauses that are included in each probe program that
needs them.  These clauses are compiled as regular probe clauses, but
they are processed separately to resolve the EPID and PRID identifiers
according to the ERROR probe rather than the probe that the clauses are
included for.

Test test/unittest/error/tst.clause_scope-begin-ended.d is new but will
fail to pass because it requires a fix in the buffer consumption code.
That will be provided in a follow-up patch/

(Special thanks to Eugene Loh for his review comments - various fixes
 he implemented got integrated into this patch.)

Orabug: 32749498
Signed-off-by: Kris Van Hees <kris.van.hees@oracle.com>
Signed-off-by: Eugene Loh <eugene.loh@oracle.com>
Reviewed-by: Kris Van Hees <kris.van.hees@oracle.com>
Reviewed-by: Eugene Loh <eugene.loh@oracle.com>