Use pcb_exitlbl in trampoline instead of allocating a label
Clauses used to be compiled into their own program, with a trampoline
prologue and epilogue surrounding the actual clause implementation.
This required a specific exit label to be allocated for use in the
trampoline because the compiler-generated label (pcb_exitlbl) was
specific to the clause implementation.

Now that clauses are compiled into proper functions, and the trampoline
is its own function (the main program function), the trampoline specific
exit label is no longer needed.  We can use the compiler-generated label
(pcb_exitlbl) instead.

Signed-off-by: Kris Van Hees <kris.van.hees@oracle.com>
Reviewed-by: Eugene Loh <eugene.loh@oracle.com>