Allow DIFOs to have a variety of flags
DIFOs could be flagged as containing destructive operations using the
specific dtdo_destructive field in dtrace_difo_t.  There is a need for
other flags to be associated with DIFOs, so the dtdo_destructive field
is being changed into a dtdo_flags field.  DTDOFLG_* constants in
difo_defines.h provide the flags that can be set and checked for.

Signed-off-by: Kris Van Hees <kris.van.hees@oracle.com>
Reviewed-by: Eugene Loh <eugene.loh@oracle.com>