Implement normalize() and denormalize()
The implementation follows the legacy implementation.  One adjustment
that was needed is to retrieve the aggregation data snapshot from
dt_normalize() to ensure that we have the latest data to work on.  This
is needed due to the on-demand data retrieval that the new aggregation
implementation uses.

Tests are simplified to specifically test the functionality they are
intended for, and some tests are added to cover gaps in the testing of
normalize() and denormalize().

Signed-off-by: Kris Van Hees <kris.van.hees@oracle.com>
Reviewed-by: Eugene Loh <eugene.loh@oracle.com>