Improvements to -xdisasm support
Add another listing, which really is the final program, with all
relocations performed and what is actually loaded into the kernel.

Add a header to all disassembly listings based on probe desc and/or
identifier name, and identifying the type of disassembly
listing.

Move disassembler defines into their own header file dt_dis.h.

Switch to using proper defines for listing types and wrap the cflags
checking together with the call to the disassembler into a macro for
each listing type.

Orabug: 32748968
Signed-off-by: Eugene Loh <eugene.loh@oracle.com>
Orabug: 32748968
Signed-off-by: Kris Van Hees <kris.van.hees@oracle.com>
Reviewed-by: Kris Van Hees <kris.van.hees@oracle.com>