PID provider implementation
This patch provides an implementation of the pid provider.  This is
a provider that makes it possible to do function boundary tracing
at the userspace level (in shared libraries and executables).

The implementation of userspace probes is built on the uprobes
support in the kernel tracing infrastructure.  The uprobes are
inode based and therefore not process specific whereas pid probes in
DTrace are process specific.  This means that there is a many-to-one
relation between pid probes and uprobes.  It is also worth noting
that a pid probe together with its underlying uprobe provides
equivalent functionality as any other non-pid DTrace probes.  The
(process specific) pid probe represents the named probe and has
D clauses associated with it.  The uprobe represents the underlying
system probe (not process specific) that our program will be attached
to.

When we want to trace the main() functoon in a process with PID 1234,
we can specify this probes as pid1234:a.out:main:entry.  A lookup is
performed to see if we already have a pid:<ino>:main:entry probe to
associate the new pid probe with.  If not, it is created.  It is this
pid:<ino>:main:entry probe that will be added to the enablings list of
probes to which we need to attach a BPF program.

Any clauses for pid1234:a.out:main:entry will be associated with that
specific probe.

When the final BPF program is constructed for a pid related enabling,
we construct the program for probe pid:<ino>:main:entry.  The program
(trampoline) will comprise basic environment setup, copying of any
possible probe arguments, and a sequence of code blocks following the
pattern:

	if (pid == <some-pid>) {
		case dt_clause_0(dctx);
		case dt_clause_1(dctx);
		<...>
		goto exit;
	} else if (pid == <some-other-pid>) {
		<...>
	}

In other words, although the uprobe will fire for any process that
reaches the probe point, we check the PID of the process that caused
the probe to fire against the registered probes and we only execute
the clauses for the matching PID.

Future patches will augment the functionality provided here with
arbitrary instruction tracing at the userspace level.

Orabug: 32749504
Signed-off-by: Kris Van Hees <kris.van.hees@oracle.com>
Reviewed-by: Eugene Loh <eugene.loh@oracle.com>