Replace probe_fini with detach, and separate detach and destroy
The probe framework provides a function dt_probe_fini() that destroys
all probes.  This function calls dt_probe_destroy() on every probe,
and it was calling the probe_fini hook for each probe to perform the
detach and destroy.  Some providers will need additional work done
between the detach and the destroy, so those need to be implemented
as separate hooks.  It is also desirable to explicitly perform the
per-probe detach as a termination of the earlier attach.

This patch removed the probe_fini hook (and related code) and instead
implements support for explicit detach and probe_destroy hooks.  The
detach must be done before probe_destroy is called.

Orabug: 32748968
Orabug: 32749504
Signed-off-by: Kris Van Hees <kris.van.hees@oracle.com>
Reviewed-by: Eugene Loh <eugene.loh@oracle.com>