Introduce a standard implementation for *_add and _del htab functions
The add and del functions are expected to follow a specific template
and can therefore be implemented by means of a generic macro.  The
probe-specific version is modified to make use of this new generic
macro.  The new macro is DEFINE_HE_STD_LINK_FUNCS.

We also introduce a DEFINE_HTAB_STD_OPS macro to can be used to
define a htab operations struct.

Orabug: 32748968
Orabug: 32749504
Signed-off-by: Kris Van Hees <kris.van.hees@oracle.com>
Reviewed-by: Eugene Loh <eugene.loh@oracle.com>