Change storage for global variables
Change how global variables are stored and accessed.  Instead of storing
them as 8-byte quantities addressed by a variable's ID, just use the value
of a single-key BPF map addressed by a variable's offset.  This allows us
to store global variables of "arbitrary" size (e.g., > 8 bytes).

A pointer to the BPF map value is kept on the stack in the DTrace context.
The get_gvar and set_gvar BPF helper functions are now obsolete.

Orabug: 32748968
Signed-off-by: Eugene Loh <eugene.loh@oracle.com>
Reviewed-by: Kris Van Hees <kris.van.hees@oracle.com>