Add a probe to test/unittest/pragma/*libdep* tests
These tests check library dependencies.  Apparently, it was
sufficient with earlier implementations to invoke "dtrace -e"
with no probes to check such dependencies, but that was not true
with early implementations of DTv2.  Further, there seems little
point in checking library dependencies if there are no probes.  In
any case, library dependencies in the absence of probes is undefined.

Add a simple probe to the libdep tests.  That way, they will not rely
on undefined behavior with regards to what library dependencies mean
in the presence of no probes.

Notes:

x Actually, it appears that depends_on is not documented at all!

x This patch should appear before the "ERROR probe implementation"
  patch, which changes the underlying behavior.  That is, "no probes"
  will nonetheless compile the ERROR probe.

Signed-off-by: Eugene Loh <eugene.loh@oracle.com>
Reviewed-by: Kris Van Hees <kris.van.hees@oracle.com>