Add support for provider-specific enabling
In general, probes are added to the list of enablings if they match
the probe description of specified in any program.  Some providers may
need to do some additional processing as part of adding a probe to the
list of enablings.  This patch adds support for an enable() provider
implementation hook.

Orabug: 32748968
Orabug: 32749504
Signed-off-by: Kris Van Hees <kris.van.hees@oracle.com>
Reviewed-by: Eugene Loh <eugene.loh@oracle.com>