Add support for ENOTSUPP error code
The ENOTSUPP error code is not meant to be exposed by the kernel, and
it is in fact specific to NFSv3.  However, the uprobes implementation
in the kernel uses it to indicate that a uprobe could not be placed at
the requested location.  This causes it to be exposed to us through
the perf_event_open() system call.

When it is encountered, the "Operation not supported" message will be
reported.

Orabug: 32748968
Orabug: 32749504
Signed-off-by: Kris Van Hees <kris.van.hees@oracle.com>
Reviewed-by: Eugene Loh <eugene.loh@oracle.com>