Fix bit-field operations
Fix bugs in the dt_cg_field_[set|get] functions.

Add tests.  Note that these tests would fail with the legacy DTrace
implementation.  The old version was incorrect since it was missing patch
"6d9c2398 Provide a more consistent implementation of the trace() action".
The difference is in results for signed bit fields with a nonzero sign bit.
Current results match GCC.

(Also note that bit fields are not packed together as efficiently as
possible.  Rather, following legacy DTrace behavior, each bit field is
aligned to the size of the next largest integer type.  This is a bug,
preventing users from reading bit fields in kernel structs, but it is
outside the scope of this patch.)

Orabug: 32748968
Signed-off-by: Eugene Loh <eugene.loh@oracle.com>
Reviewed-by: Kris Van Hees <kris.van.hees@oracle.com>