Clean up the bpflogsize tests
The bpflogsize tests are tricky, because:
- they rely on D scripts that fail the BPF verifier
- their BPF logs are too big for the default BPF log size
- their BPF logs will fit if bpflogsize is made large enough

The first question is what D script will fail the BPF verifier?  Who is
emitting bad BPF code?  The GCC cross compiler or our code generation?

The trick is to emit legimate BPF code of such complexity that the BPF
verifier cannot handle it.  The D quantize() aggregation function is a
good candidate in this regard, since it quantizes values to one of 127
different bins in highly branched code.  Repeated calls to quantize()
overwhelm the BPF verifier.

How many calls are needed depends on the underlying implementation, and
when DT_AGG_NUM_COPIES=1 was introduced the bpflogsize tests changed
behavior.

Once the number of quantize() calls is increased sufficiently, the
emitted code fails the BPF verifier, albeit with long verification
times and huge BPF log files.

It turns out that by increasing the number of quantize() calls even
further, the verifier walks branches differently and becomes overwhelmed
earlier, requiring 10x shorter verification time or log files.

Therefore, overhaul the tests:
- to make them more clear
- to induce the quicker path to BPF verification failure
- to eliminate the extended time outs

In the future, changes in the quantize() implementation or in the BPF
verifier may require further changes in these tests.

Signed-off-by: Eugene Loh <eugene.loh@oracle.com>
Reviewed-by: Kris Van Hees <kris.van.hees@oracle.com>