Change storage for local variables
Change how local variables are stored and accessed.  Instead of
storing them as 8-byte quantities on the stack addressed by variable
ID, just use the value of a single-key BPF map addressed by a variable's
offset.  This allows us to store local variables of "arbitrary" size
(e.g., > 8 bytes) and worry less about running out of stack space.
A pointer to the BPF map value is kept on the stack in the DTrace context.
In general, local-variable access now greatly mimics global-variable
access.

A dtp field dt_maxlvaralloc tracks the largest local-variable
allocation that will be needed for any pcb.

This patch also backs out patch
"a4d5f5bda15d Ensure that local variables get a 0 value upon first load",
which changed dt_cg_load_var() to load a 0 when a clause-local variable
is read but not yet initialized in that clause.  Note that:
  x Such behavior is not needed for D language compliance.
  x The old patch was motivated by a BPF verifier issue when local
    variables were kept on the stack, which is no longer the case.
  x The old patch implemented this behavior for each clause, even though
    clause-local variables should be persistent from clause to clause
    (within the same probe firing).  Thus, the old patch was faulty and
    causing test/demo/vars/clause.d to fail.

Note that this patch breaks some annotations in the disassembler, which
is fixed in a later patch.

Orabug: 32748968
Signed-off-by: Eugene Loh <eugene.loh@oracle.com>
Reviewed-by: Kris Van Hees <kris.van.hees@oracle.com>