Support safe removal of probes during dt_probe_iter()
It may be necessary to destroy probes that are found using the
dt_probe_iter() functionality.  This requires the traversal of probe
lists and hashtables to be safe in view of potential probe removal.

This patch also ensures that when a probe is destroyed, it is also
removed from the global probe array.

Orabug: 32748968
Orabug: 32749504
Signed-off-by: Kris Van Hees <kris.van.hees@oracle.com>
Reviewed-by: Eugene Loh <eugene.loh@oracle.com>