Unify common argument retrieval for providers
Multiple providers retrieve the probe arguments from the register set
and from the stack.  Function dt_cg_tramp_copy_args_from_regs() can be
called to do this in a trampoline.  The trampoline prologue will now
set up %r7 as a pointer to the mstate and %r8 as a pointer to the BPF
context.

Orabug: 32748968
Orabug: 32749504
Signed-off-by: Kris Van Hees <kris.van.hees@oracle.com>
Reviewed-by: Eugene Loh <eugene.loh@oracle.com>