Store probe id in the trampoline
The probe id (PRID) is associated with the probe and is therefore unique
for every final BPF program that is attached to a probe.  It will have
the same value for all clauses associated with that probe and it is
therefore not necessary to store it in the DTrace machine state (DMST) in
every clause.

Orabug: 32748968
Orabug: 32749504
Signed-off-by: Kris Van Hees <kris.van.hees@oracle.com>
Reviewed-by: Eugene Loh <eugene.loh@oracle.com>