Memory corruption fix during dynamic pid probe creation
The discovery and creation process for pid probes can sometimes
cause the function name to be replaced in the probe description that
is used to create pid probes.  When that probe description copies
its content from a ECB probe description, we cannot just replace the
function name value and free the original one because other code data
may still hold a reference to that function name string.

Make a copy before doing pid probe processing, and free the copy once
we are done.

Signed-off-by: Kris Van Hees <kris.van.hees@oracle.com>
Reviewed-by: Eugene Loh <eugene.loh@oracle.com>